# ARROW-7764: [C++] Don't keep a null bitmap in ArrayData if null_count == 0