This spotted a couple bugs in kernels where null bitmaps were always iterated even if null.