@pitrou 2020-02-18
@wesm I'm not entirely sure the Parquet fixes are using the right approach, advice welcome.