@pitrou 2020-02-18
Also cc @fsaintjacques for review.