@fsaintjacques 2020-02-20
I'm wary of this one, while saving space, I think it'll uncover path where we implicitly require a bitmap non-null pointer, e.g. Gandiva.