@pitrou 2020-02-20
Well, the tests pass ;-)