@wesm 2020-02-21
I'll leave this open for a bit before merging since there wasn't clear consensus...