@pitrou 2020-02-21
@wesm Well, from what I've gathered, it looks a bit messy:
* in `column_reader.cc`, there's `ColumnReaderImplBase::ReadValuesSpaced` vs. `ColumnReaderImplBase::ReadValues`
* but in the same file, there's also `TypedRecordReader::ReadValuesSpaced` vs. `TypedRecordReader::ReadValuesDense`: why? (note subtly different naming :-))

You'll see that `ReadValuesSpaced` is only called if the Parquet file may have spaced values. But perhaps in some cases the file declares an optional field that in reality is always present, in which case `null_count` ends up 0.

It seems that in `encoding.{h,cc}`, perhaps we could forego the `Decode`/`DecodeSpaced` distinction as well as `DecodeArrow`/`DecodeArrowNonNull`, and instead have the implementation switch to different paths depending on `null_count`. But that's a significant architectural change, and I don't know enough about Parquet to make a decision.