@wesm 2020-02-24
I see. I think the rationale for the Decode/DecodeSpaced distinction is that consumers of the decoded data may or may not want the spaced / Arrow-like memory layout in the decoded data, even when there are nulls. Since the spaced decoding takes extra computational work (the non-null values have to "moved into place") we don't force this semantic on public API users