@pitrou 2020-02-26
What could be done though is merge `DecodeArrow` and `DecodeArrowNonNull` together, since `DecodeArrow` already has to take a different path based on `null_count`.