@wesm 2020-02-26
@pitrou that sounds acceptable to me