@pitrou 2020-02-27
Note that this PR actually makes some decoding a bit faster:
* before:
```
BM_ArrowBinaryPlain/DecodeArrow_Dense/65536            822259 ns       822116 ns         3381 bytes_per_second=457.569M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dense/65536     778227 ns       778096 ns         4015 bytes_per_second=483.455M/s
BM_ArrowBinaryPlain/DecodeArrow_Dict/65536            3609862 ns      3609246 ns          771 bytes_per_second=104.225M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dict/65536     3500624 ns      3500022 ns          808 bytes_per_second=107.478M/s
BM_ArrowBinaryDict/DecodeArrow_Dense/65536            2702473 ns      2702013 ns         1035 bytes_per_second=139.22M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dense/65536     2689046 ns      2688489 ns         1034 bytes_per_second=139.921M/s
BM_ArrowBinaryDict/DecodeArrow_Dict/65536             4219876 ns      4219148 ns          670 bytes_per_second=89.159M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dict/65536      4143485 ns      4142749 ns          677 bytes_per_second=90.8032M/s
```
* after:
```
BM_ArrowBinaryPlain/DecodeArrow_Dense/65536            690074 ns       689951 ns         4069 bytes_per_second=545.22M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dense/65536     692913 ns       692758 ns         4046 bytes_per_second=543.01M/s
BM_ArrowBinaryPlain/DecodeArrow_Dict/65536            3275331 ns      3274655 ns          864 bytes_per_second=114.875M/s
BM_ArrowBinaryPlain/DecodeArrowNonNull_Dict/65536     3263202 ns      3262458 ns          852 bytes_per_second=115.304M/s
BM_ArrowBinaryDict/DecodeArrow_Dense/65536            2476155 ns      2475691 ns         1128 bytes_per_second=151.947M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dense/65536     2612045 ns      2611368 ns         1068 bytes_per_second=144.053M/s
BM_ArrowBinaryDict/DecodeArrow_Dict/65536             3986622 ns      3985459 ns          705 bytes_per_second=94.3868M/s
BM_ArrowBinaryDict/DecodeArrowNonNull_Dict/65536      3973003 ns      3971878 ns          705 bytes_per_second=94.7096M/s
```