@pitrou 2020-02-27
AppVeyor failure is due to a conda-forge package regression... :-/