@emkornfield 2020-02-27
Sorry I missed the initial ping.  I can take a look tonight.  If the question was about parquet read path I have no opinion yet